+++ NewEva/Model/Customer/Organization.cs
@@ -10,13 +10,13 @@ namespace NewEva.Model
         public string NameFull { get; set; } //Полное наименование
         public string NameShort { get; set; } //Сокращенное наименование
         public string OrganizationForm { get; set; } //Организационно-правовая форма
-        public int OGRN { get; set; } //ОГРН
+        public long OGRN { get; set; } //ОГРН
         public DateTime? DateRegistration { get; set; } //Дата регистрации
         public int INN { get; set; } //ИНН
         public int KPP { get; set; } //КПП
         public string NameBank { get; set; } //Название банка
         public int BIK { get; set; } //БИК Банка
-        public int PayAccount { get; set; } //Расчетный счет
-        public int CorrAccount { get; set; } //Корреспондентский счет
+        public long PayAccount { get; set; } //Расчетный счет
+        public long CorrAccount { get; set; } //Корреспондентский счет
     }
 }
